requirements.txt
Flask==3.1.2
flask-cors==5.0.0
Flask-SQLAlchemy==3.1.1
gunicorn==21.2.0
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.0
PyPDF2==3.0.1
python-docx==1.1.0
pytesseract==0.3.10
Pillow>=10.0.0
playwright==1.40.0
trafilatura==1.6.3
psycopg2-binary==2.9.9
